try refactoring scan page integration test
