+++ app/test_driver/scan_page/encointer-receive-qr-1.jpg
