+++ app/test_driver/scan_page_test.dart
@@ -7,28 +7,23 @@ import 'package:test/test.dart';
 import 'helpers/helper.dart';
 
 void main() {
-  FlutterDriver? driver;
+  late FlutterDriver driver;
+  const receiveQrImagePath = 'test_driver/scan_page/encointer-receive-qr-1.jpg';
 
-  group('scan-page', () {
-    setUpAll(() async {
-      driver = await FlutterDriver.connect();
-      // waits until the firs frame after ft startup stabilized
-      await driver!.waitUntilFirstFrameRasterized();
-    });
-
-    test('scan-page-screenshot', () async {
-      final file = File('test_driver/resources/encointer-receive-qr-1.jpg');
-      final bytes = await file.readAsBytes();
-      final base64 = base64Encode(bytes);
-
-      // set the background in the MockScanPage
-      await driver!.requestData(base64);
-
-      await driver!.takeScreenshot('mock-scan-receive');
-    });
+  setUpAll(() async {
+    driver = await FlutterDriver.connect();
+    await driver.waitUntilFirstFrameRasterized();
   });
 
-  tearDownAll(() async {
-    if (driver != null) await driver!.close();
+  test('scan-page-screenshot', () async {
+    final file = File(receiveQrImagePath);
+    final bytes = await file.readAsBytes();
+    final base64 = base64Encode(bytes);
+
+    // set the background in the MockScanPage
+    await driver.requestData(base64);
+    await driver.takeScreenshot('mock-scan-receive');
   });
+
+  tearDownAll(() async => driver.close());
 }
